+++ CHANGELOG.rst
@@ -16,6 +16,8 @@ Added
 - **Projectroles**
     - Projectroles app settings (#532)
     - Remote sync for projectroles app setting (#533)
+    - IP address based access restriction for projects (#531)
+    - ``is_delegate()`` and ``is_owner_or_delegate()`` helpers for ``Project`` model
 
 Changed
 -------